requirements.txt
accelerate==0.24.1
autocommand==2.2.2
backports.tarfile==1.2.0
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
diffusers==0.30.2
filelock==3.13.1
fsspec==2024.6.1
huggingface-hub==0.33.0
idna==3.10
importlib-metadata==8.7.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jinja2==3.1.4
MarkupSafe==2.1.5
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==25.0
peft==0.8.2
pillow==11.0.0
pip==25.1.1
platformdirs==4.2.2
psutil==7.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==80.9.0
sympy==1.13.3
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.67.1
transformers==4.38.2
typeguard==4.3.0
typing-extensions==4.12.2
urllib3==2.5.0
wheel==0.45.1
zipp==3.23.0
